refactor: move domain features into modules
