+++ apps/frontend/src/app/app.routes.ts
@@ -8,13 +8,13 @@ import { NewPasswordPage } from './auth/new-password-page';
 import { ResetPasswordPage } from './auth/reset-password-page';
 import { SignInPage } from './auth/signin-page';
 import { SignUpPage } from './auth/signup-page';
-import { DonorsGrid } from './components/donors/donors-grid';
-import { HouseholdDetail } from './components/households/household-detail';
-import { HouseholdsGrid } from './components/households/households-grid';
-import { PersonDetail } from './components/persons/person-detail';
-import { PersonsGrid } from './components/persons/persons-grid';
-import { TagsGridComponent } from './components/tags/tags-grid';
-import { VolunteersGrid } from './components/volunteers/volunteers-grid';
+import { DonorsGrid } from './features/donors/ui/donors-grid';
+import { HouseholdDetail } from './features/households/ui/household-detail';
+import { HouseholdsGrid } from './features/households/ui/households-grid';
+import { PersonDetail } from './features/persons/ui/person-detail';
+import { PersonsGrid } from './features/persons/ui/persons-grid';
+import { TagsGridComponent } from './features/tags/ui/tags-grid';
+import { VolunteersGrid } from './features/volunteers/ui/volunteers-grid';
 import { Dashboard } from './layout/dashboards/dashboard';
 import { Summary } from './temp/summary';
