+++ apps/frontend/src/app/features/households/ui/household-detail.html
